refactor: extract shell LLM context gathering into shell_context.py
Move gather_llm_context(), redact_for_llm(), and _detect_shell_tools()
from shell_commands.py into a new shell_context.py so that shell_capture.py
can import them without creating a shell_capture → shell_commands circular
dependency. shell_commands.py re-exports the names for callers that haven't
migrated their imports yet.
